Re-work EffectLib into a multi-module project.
This was kind of a huge mess. The annoying problem here was the change to require using the new "force display" flag in the Particle.display() method with a new signature.

In order to use that, we have to build against a newer version of the Spigot API.

However, for whatever reason, the Sound class was changed from an enum to an interface just prior to the version that added the new display signature.

And of course, building against the new Sound interface made the plugin not backwards-compatible.

So, making the project multi-module was the only solution I could think of.

I considered just dropping backwards compatibility, but due to the proximity of these two breaking changes I would have to resign to EffectLib only working on 1.21.4 and above, which did not seem ok and would not even work for my own uses, the Magic plugin which now only supports 1.20.5 and up.